+++ CHANGELOG.md
@@ -17,7 +17,7 @@
 - All (most) types are now living in their own crate `twitch_types`
 - Features for clients are now named after the client, e.g feature `reqwest_client` is now simply `reqwest`
 - Fixed wrong type on `UserAuthorizationGrantV1::client_id`
-- Deprecate `bits_voting_enabled`, `bits_per_vote` and `bits_votes` on `Create Poll` and `Poll` and made the fields optional
+- Removed `bits_voting_enabled`, `bits_per_vote` and `bits_votes` on `Create Poll` and `Poll`
 - `Get Clips` takes and returns `ClipId` instead of a string
 - Made `GetCustomRewardRedemptionRequest::reward_id` optional
 - `typed-builder` is now optional and not enabled by default, enable feature `typed-builder` to use the `::builder()` methods